requirements.txt
black==19.10b0
Pillow==6.2.1
numpy==1.17.4